[OPENENGSB-1786] Move workflow tests from itest to core.workflow
