+++ core/workflow/src/test/java/org/openengsb/core/workflow/persistence/PersistenceRuleManagerTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.openengsb.itests.workflow;
+package org.openengsb.core.workflow.persistence;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.not;
@@ -35,7 +35,7 @@ import org.openengsb.core.api.workflow.RuleBaseException;
 import org.openengsb.core.api.workflow.model.RuleBaseElementId;
 import org.openengsb.core.api.workflow.model.RuleBaseElementType;
 
-public class PersistenceRuleManagerIT extends AbstractRuleManagerIT {
+public class PersistenceRuleManagerTest extends AbstractRuleManagerTest {
 
     @Test
     public void testGetRuleBase() throws Exception {
@@ -61,7 +61,7 @@ public class PersistenceRuleManagerIT extends AbstractRuleManagerIT {
     public void testAddRule() throws Exception {
         RuleBaseElementId id = new RuleBaseElementId(RuleBaseElementType.Rule, "org.openengsb", "test3");
         ruleManager.add(id, "when\n" + "  e : Event()\n" + "then\n"
-                + "  example.doSomething(\"this rule was added by the addrule-function\");\n");
+                + "  example2.doSomething(\"this rule was added by the addrule-function\");\n");
 
         createSession();
         executeTestSession();
@@ -101,7 +101,7 @@ public class PersistenceRuleManagerIT extends AbstractRuleManagerIT {
     public void testAddGlobal() throws Exception {
         ruleManager.addGlobal("java.util.Random", "bla");
         ruleManager.add(new RuleBaseElementId(RuleBaseElementType.Rule, "bla"),
-            "when\n then example.doSomething(\"\" + bla.nextInt());");
+            "when\n then example2.doSomething(\"\" + bla.nextInt());");
         createSession();
         session.setGlobal("bla", new Random());
         session.insert(new Event());
@@ -165,7 +165,7 @@ public class PersistenceRuleManagerIT extends AbstractRuleManagerIT {
     @Test
     public void testAddOtherPackages() throws Exception {
         RuleBaseElementId id = new RuleBaseElementId(RuleBaseElementType.Rule, "at.ac.tuwien", "hello42");
-        ruleManager.add(id, "when\nthen\nexample.doSomething(\"bla\");");
+        ruleManager.add(id, "when\nthen\nexample2.doSomething(\"bla\");");
         createSession();
         executeTestSession();
         assertTrue(listener.haveRulesFired("at.ac.tuwien.hello42"));
@@ -174,9 +174,9 @@ public class PersistenceRuleManagerIT extends AbstractRuleManagerIT {
     @Test
     public void testRulesInDifferentPackages() throws Exception {
         RuleBaseElementId id = new RuleBaseElementId(RuleBaseElementType.Rule, "at.ac.tuwien", "hello42");
-        ruleManager.add(id, "when\nthen\nexample.doSomething(\"bla\");");
+        ruleManager.add(id, "when\nthen\nexample2.doSomething(\"bla\");");
         id = new RuleBaseElementId(RuleBaseElementType.Rule, "org.openengsb", "hello42");
-        ruleManager.add(id, "when\nthen\nexample.doSomething(\"bla\");");
+        ruleManager.add(id, "when\nthen\nexample2.doSomething(\"bla\");");
         createSession();
         executeTestSession();
         assertTrue(listener.haveRulesFired("org.openengsb.hello42", "at.ac.tuwien.hello42"));
