+++ core/workflow/src/test/java/org/openengsb/core/workflow/persistence/PersistenceTestUtil.java
@@ -17,11 +17,30 @@
 
 package org.openengsb.core.workflow.persistence;
 
+import java.io.File;
+import java.io.IOException;
+import java.net.URL;
+import java.util.List;
+
+import org.apache.commons.io.FileUtils;
+import org.mockito.Mockito;
+import org.openengsb.core.api.Event;
+import org.openengsb.core.api.persistence.PersistenceException;
+import org.openengsb.core.api.persistence.PersistenceService;
+import org.openengsb.core.api.workflow.RuleBaseException;
 import org.openengsb.core.api.workflow.RuleManager;
+import org.openengsb.core.persistence.internal.NeodatisPersistenceService;
 import org.openengsb.core.test.DummyPersistence;
 import org.openengsb.core.workflow.internal.persistence.PersistenceRuleManager;
+import org.openengsb.core.workflow.model.GlobalDeclaration;
+import org.openengsb.core.workflow.model.ImportDeclaration;
+import org.osgi.framework.Bundle;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public final class PersistenceTestUtil {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(PersistenceTestUtil.class);
 
     public static RuleManager getRuleManagerWithMockedPersistence() throws Exception {
         PersistenceRuleManager manager = new PersistenceRuleManager();
@@ -33,6 +52,64 @@ public final class PersistenceTestUtil {
 
     public static RuleManager getRuleManager() throws Exception {
         return getRuleManagerWithMockedPersistence();
+    }
+
+    public static RuleManager getRuleManagerWithPersistenceService() throws PersistenceException, IOException,
+        RuleBaseException {
+        PersistenceRuleManager manager = new PersistenceRuleManager();
+        NeodatisPersistenceService persistence = createPersistence();
+        manager.setPersistence(persistence);
+        manager.init();
+        return manager;
+    }
+
+    private static NeodatisPersistenceService createPersistence() throws PersistenceException, IOException {
+        LOGGER.debug("creating persistence");
+        final File dataFile = new File("data");
+        FileUtils.deleteQuietly(dataFile);
+        File refData = new File("data.ref");
+        if (!refData.exists()) {
+            LOGGER.debug("creating reference persistence");
+            createReferencePersistence();
+        }
+        FileUtils.copyFile(refData, dataFile);
+        NeodatisPersistenceService persistence = new NeodatisPersistenceService("data", Mockito.mock(Bundle.class));
+        return persistence;
+    }
+
+    public static void createReferencePersistence() throws PersistenceException, IOException {
+        FileUtils.deleteQuietly(new File("data.ref"));
+        NeodatisPersistenceService persistence = new NeodatisPersistenceService("data.ref", Mockito.mock(Bundle.class));
+        persistence.create(new ImportDeclaration(Event.class.getName()));
+        readImports(persistence);
+        readGlobals(persistence);
+    }
+
+    private static void readGlobals(PersistenceService persistence) throws IOException, PersistenceException {
+        URL globalURL = ClassLoader.getSystemResource("rulebase/globals");
+        File globalFile = FileUtils.toFile(globalURL);
+        List<String> globalLines = FileUtils.readLines(globalFile);
+        for (String s : globalLines) {
+            String[] parts = s.split(" ");
+            persistence.create(new GlobalDeclaration(parts[0], parts[1]));
+        }
+    }
+
+    private static void readImports(PersistenceService persistence) throws IOException, PersistenceException {
+        URL importsURL = ClassLoader.getSystemResource("rulebase/imports");
+        File importsFile = FileUtils.toFile(importsURL);
+        List<String> importLines = FileUtils.readLines(importsFile);
+        for (String s : importLines) {
+            persistence.create(new ImportDeclaration(s));
+        }
+    }
+
+    public static void cleanup() {
+        FileUtils.deleteQuietly(new File("data"));
+    }
+
+    public static void cleanupReferenceData() {
+        FileUtils.deleteQuietly(new File("data.ref"));
     }
 
     private PersistenceTestUtil() {
