+++ core/workflow/src/test/java/org/openengsb/core/workflow/persistence/RuleListener.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.openengsb.itests.workflow;
+package org.openengsb.core.workflow.persistence;
 
 import java.util.Arrays;
 import java.util.HashSet;
