+++ core/workflow/src/test/resources/rulebase/globals
@@ -6,3 +6,4 @@ org.openengsb.core.workflow.DummyDeploy deploy
 org.openengsb.core.workflow.DummyReport report
 org.openengsb.core.workflow.DummyIssue issue
 org.openengsb.core.workflow.DummyService myservice
+org.openengsb.domain.example.ExampleDomain example2
